services/resource-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
pydantic[email]==2.5.2
email-validator==2.1.0
pydantic-settings==2.1.0
httpx==0.25.2
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0
